Adding AbstracApplicationAuthzTestCase for IDENTITY-5580 (#1239)
Testcase for IDENTITY-5580
