+++ pom.xml
@@ -1533,7 +1533,7 @@
         <carbon.identity.auth.version>1.1.1</carbon.identity.auth.version>
         <identity.event.handler.account.lock.version>1.1.0</identity.event.handler.account.lock.version>
         <identity.event.handler.notification.version>1.0.0</identity.event.handler.notification.version>
-        <identity.app.authz.xacml.version>1.0.0</identity.app.authz.xacml.version>
+        <identity.app.authz.xacml.version>1.0.1</identity.app.authz.xacml.version>
         <!--<identity.agent.entitlement.proxy.version>5.1.1</identity.agent.entitlement.proxy.version>-->
         <!--<identity.carbon.auth.signedjwt.version>5.1.1</identity.carbon.auth.signedjwt.version>-->
         <!--<identity.userstore.cassandra.version>5.1.1</identity.userstore.cassandra.version>-->
